feat:默认重建output目录

+++ internal/compile/compile.go
@@ -28,23 +28,23 @@ func NewCompiler(cfg *conf.Config) *Compiler {
 	}
 }
 
-func (c *Compiler) Compile(desc proto.DescriptorSource) (err error) {
+func (c *Compiler) Compile(desc proto.DescriptorSource, deleteDirectory bool) (err error) {
 
 	for _, itr := range c.config.Generate.Plugins {
 		metaCmds := c.generateCmd(desc, itr.Type)
 		//插件
 		arg := fmt.Sprintf("--%s_out=", itr.Name)
 		outputPath := os.ExpandEnv(itr.Output)
-		/*
+		if deleteDirectory {
 			err = os.RemoveAll(outputPath)
 			if err != nil {
-				log.Println("remove path failed ", outputPath)
+				log.Fatalf("remove path failed ", outputPath)
 			}
-		*/
+		}
+
 		err = os.MkdirAll(outputPath, os.ModePerm)
 		if err != nil {
 			log.Fatalf("MkdirAll failed [%v]", err)
-			break
 		}
 		log.Println("out path ", outputPath)
 		//参数
